Updates to BinaryTree lib and tests
Added more implementation to BinaryTree since... there was none, ha. I
also made BinaryTree a subclass of Tree. Not sure if it will really be
useful but it lets us use some of the inhertitence rules in C++ so we
can get used to using it.

Added some extra spacing in the sorting test to seperate some of the
test output.