Implemented Running Tests From Make Target: #2
Top-level makefile now has a `run-test` target. Resolves issue #2. When
compiling the test executables, I place them in the top-level bin folder
and this is where I run the tests from. Libs are also placed in this folder
but I only run the files that are executable.

This new target is not included in the default make target since I like
my top-level `make` to only compile code, create docs, etc, not actually
run anything. So, if you want to run tests, just do `make run-test`.