Work on #1: Dependencies and Build Order
Been trying to get dependecies resolved and build order. All hpp
templated headers are now correctly linked but my static libraries
are still not linked in correctly. If I put the implmentatiaon in
the header, I'm golden. But when I move it to the impl .cpp file,
I get the notorious 'undefined reference to...' error. I've tried
every which was I can think of to order the build rules but nothing
works. It was working before but I didn't quite understand every-
thing so I started from scratch. Looks like it didn't work...'