Fixed Static Library Build Issue & Other Updates
Finally figured out why functions in my static libraries were not being
recognized at the linking stage. It was because I didn't add the name-
space to the implementation of the function... Which doesn't make a bit
of sense since I did `using namespace searchUtil;` which should have
made it so that I didn't have to scope the function AGAIN. However,
this was not the case. Nonetheless... it works now so I'm happy. :)

Updated what the makefile outputs when it is deleting directories.
There is now a `done` that shows up once is has been completed so
that the user knows if it has been completed or not, like people
do in the typical config scripts.