Resolved Library Scoping Issue
In the library, I was having issues with namespaces. In my impl file,
some of my functions were not being recognized at the linking step. It
turns out the problem was that I had to use the same syntax in the impl
file for namespaces as I did in the header. So I just wrapped everything
in the impl file in a `namespace ssearchUtil { ... }` and it worked!
Using `using namespace searchUtil;` did not work for some reason in the
impl file for non-member, non-inline functions. Not sure why yet...