Added New Library StringUtil
Created a new library: StringUtil. This library should grow over time
when and where we find useful functions that should be in the in here
for strings. I added a basic templated function toString that will
convert any type (has to have operator<< defined) to a string using
stringstreams.

Also added in a length function that just returns the length of a
string. I added this in so that I could have a working impl file
as well.

Added more tests and took out junk code from Searching lib.