Created install make target (Work on #8).
Created the make target `install`. All if does right now is take the
*.a's from the bin dir and COPY them to /usr/local/lib. However, I was
incorrect about the permissions. You still have to have sudo access to
put files in /usr/local/**/, lame. I suppose we will just have to say
"install this using sudo". We could add a note about installing in
their home directory and then updating the make rules to include stuff
in the home directory.

The biggest thing that needs to be done with this is to get the headers
into the /usr/local/include folder some way with some sort of unique
name. Work on that later.