Created qa dir with test build script.
Created new directory for QA type stuff (might want to think about
moving some of the other stuff into this folder like docs or maybe
even tests? Not sure.

First pass at a test-build script. Takes all output and funnels to
a logfile that we can email or send to CDash if we ever use it. We can
just cron the script in on some machine for now to do nightly, weekly,
or monthly runs. Doesn't matter to me. I just wanted something in place
that we could setup as a cronjob for now.